+++ src/test/java/de/alpharogroup/evaluate/object/verifier/ContractVerifierTest.java
@@ -1,27 +1,43 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package de.alpharogroup.evaluate.object.verifier;
-
-import java.util.function.Function;
 
 import org.testng.annotations.Test;
 
-import de.alpharogroup.evaluate.object.BaseEnhancedRandomExtensions;
 import de.alpharogroup.test.objects.Member;
 import de.alpharogroup.test.objects.Person;
 
+/**
+ * The unit test class for the class {@link ContractVerifier}
+ */
 public class ContractVerifierTest
 {
 
+	/**
+	 * Test method for {@link ContractVerifier#verify()}
+	 */
 	@Test
 	public void testOfVerify()
 	{
-		Function<Class<Person>, Person> factoryFunction = clazz -> Person.builder()
-			.name(BaseEnhancedRandomExtensions.nextObject(String.class))
-			.about(BaseEnhancedRandomExtensions.nextObject(String.class))
-			.married(BaseEnhancedRandomExtensions.nextObject(Boolean.class))
-			.nickname(BaseEnhancedRandomExtensions.nextObject(String.class)).build();
-		ContractVerifier.of(Person.class)
-		.withFactoryFunction(factoryFunction)
-		.verify();
+		ContractVerifier.of(Person.class).verify();
 		ContractVerifier.of(Member.class).verify();
 	}
 }